fix systemd nginx interop
- drop the .socket unit file - this didn't work as expected
- fix up the .service unit file
- add systemd notify support

The service was attemting to start multiple times because it was enabled
by default and the socket also triggered it. The systemd socket also
caused an error since it used the same address/port. Also nginx would
fail to start because the aiohttp server wasn't ready soon enough, so
using systemd notify lets us delay nginx starting until we are ready.